+++ examples/tests/test_buffer_array.py
@@ -6,6 +6,7 @@ import api
 def test_create_empty_buffer():
     name = "drum"
     duration_ms = 500
+    # TODO: check if a buffer already exists
     buf = api.create_empty_buffer(name, duration_ms)
     api.post(f"created buffer name: '{name}' duration(ms): '{duration_ms}'")
     api.post(f"framecount: {buf.framecount}")
@@ -55,3 +56,17 @@ def test_buffer_set_samples():
     # xs = array("d", [math.sin(i) for i in t])
     buf.set_samples(xs)
     api.post(f"set {n_samples} samples to buffer {name}")
+
+def test_buffer_set_samples2():
+    name = "drum1"
+    duration_ms = 500
+    buf = api.create_empty_buffer(name, duration_ms)
+    api.post(f"testing auto-resizing of buffer size")
+    api.post("buffer.set_sample example with numpy and scipy.signal")
+    api.post(f"created buffer name: '{name}' duration(ms): '{duration_ms}'")
+    api.post(f"framecount: {buf.framecount}")
+    n_samples = 2 * buf.n_samples
+    t = linspace(0, 1, n_samples, endpoint=False)
+    xs = array("d", [math.sin(i * 2 * math.pi * 5) for i in t])
+    buf.set_samples(xs)
+    api.post(f"set {buf.n_samples} samples to buffer {name}")