+++ examples/tests/test_buffer_np.py
@@ -35,3 +35,18 @@ def test_buffer_set_samples():
     xs = signal.sawtooth(2 * np.pi * 5 * t)
     buf.set_samples(xs)
     api.post(f"set {n_samples} samples to buffer {name}")
+
+def test_buffer_set_samples2():
+    name = "drum1"
+    duration_ms = 500
+    buf = api.create_empty_buffer(name, duration_ms)
+    api.post(f"testing auto-resizing of buffer size")
+    api.post("buffer.set_sample example with numpy and scipy.signal")
+    api.post(f"created buffer name: '{name}' duration(ms): '{duration_ms}'")
+    api.post(f"framecount: {buf.framecount}")
+    n_samples = buf.n_samples * 2
+    t = np.linspace(0, 1, n_samples, endpoint=False, dtype=np.float64)
+    xs = signal.sawtooth(2 * np.pi * 5 * t)
+    buf.set_samples(xs)
+    api.post(f"set {buf.n_samples} samples to buffer {name}")
+